email-service/requirements.txt
Flask==1.1.2  
SQLAlchemy==1.3.16 
flask-marshmallow==0.12.0 
Flask-SQLAlchemy==2.4.1  
marshmallow-sqlalchemy==0.23.0 
python-dotenv==0.13.0 
pymysql==0.9.3
gunicorn==20.0.4
google-cloud-tasks==1.5.0
googleapis-common-protos==1.51.0
flask-mail==0.9.1